Move generic component
